src/requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-crispy-forms==1.7.2
gunicorn==19.9.0
image==1.5.27
ImageHash==4.0
numpy==1.16.1
Pillow==5.4.1
pytz==2018.9
PyWavelets==1.0.1
scipy==1.2.1
six==1.12.0
